app/requirements.txt
alabaster==0.7.10
amqp==2.2.2
anyjson==0.3.3
asn1crypto==0.23.0
Babel==2.5.1
backports.ssl==0.0.9
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
boto==2.48.0
boto3==1.4.7
botocore==1.7.34
bumpversion==0.5.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
confusable-homoglyphs==2.0.2
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.1
cssselect==1.0.1
ddtrace==0.10.0
dj-stripe==1.0.0.post1
Django==1.11.6
django-braces==1.11.0
django-cachalot==1.5.0
django-cors-headers==2.1.0
django-countries==5.0
django-debug-toolbar==1.8
django-extensions==1.9.6
django-filter==1.1.0
django-model-utils==3.0.0
django-money==0.12
django-polymorphic==1.3
django-redis==4.8.0
django-redis-cache==1.7.1
django-registration==2.3
django-rest-swagger==2.1.2
django-sendgrid-v5==0.6.7
django-ses==0.8.3.1
django-social-auth==0.7.28
django-storages==1.6.5
django-taggit==0.22.1
django-taggit-serializer==0.1.5
djangorestframework==3.7.1
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
docutils==0.14
drf-extensions==0.3.1
drf-ujson==1.2
ecdsa==0.13
elasticsearch==5.4.0
enum34==1.1.6
et-xmlfile==1.0.1
Fabric==1.14.0
fabric-gce-tools==0.6.1
feedfinder2==0.0.4
feedparser==5.2.1
FullContact.py==0.0.6
future==0.16.0
futures==3.1.1
gevent==1.2.2
greenlet==0.4.12
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
importlib==1.0.4
inflection==0.3.1
ipaddress==1.0.18
itypes==1.1.0
jdcal==1.3
jieba==0.39
Jinja2==2.9.6
jmespath==0.9.3
jsonfield==2.0.2
kombu==4.1.0
lxml==4.1.0
MarkupSafe==1.0
msgpack-python==0.4.8
ndg-httpsclient==0.4.3
nylas==3.0.0
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.4.9
paramiko==2.3.1
Pillow==2.5.1
psycogreen==1.0
psycopg2==2.7.3.1
py-moneyed==0.7.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
pycrypto==2.6.1
pycryptodomex==3.4.7
Pygments==2.2.0
pyjwkest==1.4.0
PyJWT==1.5.3
PyNaCl==1.1.2
pyOpenSSL==17.3.0
python-dateutil==2.4.0
python-doc-inherit==0.3.0
python-http-client==3.0.0
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2017.2
PyYAML==3.11
raven==6.2.1
redis==2.10.6
requests==2.18.4
requests-file==1.4.2
requests-oauthlib==0.8.0
requirements-dev.txt==0.0.7
rsa==3.4.2
s3transfer==0.1.11
sendgrid==5.2.1
simplejson==3.11.1
six==1.11.0
smtpapi==0.3.1
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.4.0
Sphinx==1.6.5
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
stripe==1.70.0
tldextract==1.5.1
tqdm==4.19.4
typing==3.6.2
ujson==1.35
uritemplate==3.0.0
urllib3==1.22
URLObject==2.4.3
vine==1.1.4
wrapt==1.10.11
